Added and fixed #13 and #7

+++ src/events/main/messageCreate.ts
@@ -4,28 +4,30 @@ import IEvent from "../event";
 import { prefix } from "../../configs/config.json"
 import { PinHandler } from "../../server/server_PinHandler";
 import getWords from "../../words/word_index";
-import { Player } from "discord-music-player";
+import { Player } from "@jadestudios/discord-music-player";
 
 export default class MessageCreate implements IEvent {
 	public readonly name: string = 'messageCreate';
 	public readonly once: boolean = false;
 	private commands = getCommands();
 	private words = getWords();
+	private escapeRegex = (str: string) => str.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+
 	public execute(...args: any[]): void {
 		const message = <Message>args[0]; //Since messageCreate only has one argument
-		if (message.member == null || message.guild == null || message.client.user == null) return;
-		if (message.channel.type === "DM") return;
+		if (message.client.user == null || message.channel.type === "DM" || message.guild == null || message.member == null || message.author.bot) return;
+		const prefixRegex = new RegExp(`^(<@!?${message.client.user.id}>|${this.escapeRegex(prefix)})\\s*`);
 
-		if (message.type === 'CHANNEL_PINNED_MESSAGE') { //Makes it so that new PinHandler is made during pinned checks
-			const pinHandler = new PinHandler(`${message.guild.id}.db`);
+		if (message.guild.me?.isCommunicationDisabled()) return; //No response during timeout
+
+		if (message.type === 'CHANNEL_PINNED_MESSAGE') { //Check for new pins
+			const pinHandler = new PinHandler(`${message.guild.id}.db`); //Makes it so that new PinHandler is made during pinned checks
 			if (pinHandler.hasArchiveChannel()) {
 				message.delete();
-			}		
+			}
 			return;
 		}
 
-		const escapeRegex = (str: string) => str.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-		const prefixRegex = new RegExp(`^(<@!?${message.client.user.id}>|${escapeRegex(prefix)})\\s*`);
 		const player = args[1] as Player; //args[1] should be a dmp - player object
 
 		if (!prefixRegex.test(message.content)) {
@@ -37,7 +39,6 @@ export default class MessageCreate implements IEvent {
 					} catch (error) {
 						console.error(error);
 					}
-					
 				}
 			}
 			return;
@@ -51,7 +52,7 @@ export default class MessageCreate implements IEvent {
 		if (!this.commands.has(command)) return;
 
 		try {
-			this.commands.get(command)!.execute(prefix, command, message, messageArgs, [player]); 
+			this.commands.get(command)!.execute(prefix, command, message, messageArgs, [player]);
 		} catch (error) {
 			console.error(error);
 			message.reply('Failed to execute command');
